refactor: delete project settings loader middleweare
